Separate enums that are not related to the machine
