+++ src/machines/reviewTask/enums.ts
@@ -27,28 +27,4 @@ enum CompoundStates {
     draft = "draft",
 }
 
-enum ClassificationEnum {
-    "not-fact" = 0,
-    "false",
-    "misleading",
-    "unsustainable",
-    "unverifiable",
-    "exaggerated",
-    "arguable",
-    "trustworthy-but",
-    "trustworthy",
-}
-
-enum Roles {
-    Regular = "regular",
-    FactChecker = "fact-checker",
-    Admin = "admin",
-}
-
-export {
-    ReviewTaskEvents,
-    ReviewTaskStates,
-    CompoundStates,
-    ClassificationEnum,
-    Roles,
-};
+export { ReviewTaskEvents, ReviewTaskStates, CompoundStates };
