+++ src/store/types.ts
@@ -1,4 +1,5 @@
 import { WidthBreakpoints } from "../hooks/useMediaQueryBreakpoints";
+import { Roles } from "../types/enums";
 
 export enum ActionTypes {
     TOGGLE_MENU,
@@ -33,7 +34,7 @@ export interface RootState {
     };
     login: boolean;
     autoSave: boolean;
-    role: string;
+    role: Roles;
     vw: WidthBreakpoints;
     selectedDataHash: string;
     selectedPersonality: any;
