+++ src/types/enums.ts
@@ -1,0 +1,19 @@
+enum ClassificationEnum {
+    "not-fact" = 0,
+    "false",
+    "misleading",
+    "unsustainable",
+    "unverifiable",
+    "exaggerated",
+    "arguable",
+    "trustworthy-but",
+    "trustworthy",
+}
+
+enum Roles {
+    Regular = "regular",
+    FactChecker = "fact-checker",
+    Admin = "admin",
+}
+
+export { ClassificationEnum, Roles };
